+++ Marlin/src/HAL/LPC1768/fast_pwm.cpp
@@ -30,7 +30,7 @@ void set_pwm_duty(const pin_t pin, const uint16_t v, const uint16_t v_size/*=255
     LPC176x::pwm_write_ratio(pin, invert ? 1.0f - (float)v / v_size : (float)v / v_size);  // map 1-254 onto PWM range
 }
 
-void set_pwm_frequency(const pin_t pin, const int f_desired) {
+void set_pwm_frequency(const pin_t pin, const uint16_t f_desired) {
   LPC176x::pwm_set_frequency(pin, f_desired);
 }
